server-after/catalog/src/catalog-api/requirements.txt
Eve==2.1.0
Flask-Cors==3.0.10
requests==2.28.1
hypermea-core==0.9.25
# start: added by add-validation
isodate==0.6.1
# end: added by add-validation
